requirements.txt
dspy-ai>=1.0
python-dotenv>=1.0
ollama>=0.5.1
mlflow>=2.0
firecrawl-py>=0.1.0
numpy>=1.0
pytest>=7.0
pytest-asyncio>=0.21
rich>=13.0.0
termcolor>=2.0.0
tenacity>=8.2.0
textual>=0.50
simpledspy>=0.3.1
feedparser>=6.0.10
PyYAML>=6.0
requests>=2.31.0
toml>=0.10.2
pydantic>=2.0.0
